fix(fused_linear_jsd): project logits directly in FP32
The forward projected hidden states with a low-precision GEMM and then
cast the rounded result to FP32, discarding the FP32 accumulator cuBLAS
already computed. JSD gradients difference near-equal distributions, so
the rounding cancels catastrophically (4-23% grad error in bf16).

Use torch.mm(..., out_dtype=torch.float32) where supported (torch>=2.8,
CUDA sm_80+, fp16/bf16) to keep the accumulator at no cost, with an
explicit FP32 GEMM fallback elsewhere. Apply the same FP32 projection to
the Ascend and cuTile forwards. Fix the test oracle, which shared the
same cast-after-matmul defect, and switch its degenerate rand inputs to
randn so the bug is no longer masked.

+++ src/liger_kernel/ops/backends/_ascend/ops/fused_linear_jsd.py
@@ -90,10 +90,12 @@ def fused_linear_jsd_forward(
         teacher_input_chunk = teacher_input[start_idx:end_idx]
 
         # shape: chunk_size x V
-        # For anything starting from logits to the final JSD loss, we do computation
-        # in FP32 to avoid losing numerical stability.
-        student_logits_chunk = (student_input_chunk @ student_weight.t()).to(torch.float32)
-        teacher_logits_chunk = (teacher_input_chunk @ teacher_weight.t()).to(torch.float32)
+        # Project directly in FP32: the GEMM accumulator is FP32, so keep it
+        # instead of rounding logits to the input dtype first. Numerically
+        # equivalent to torch.mm(..., out_dtype=torch.float32) where supported;
+        # NPU has no out_dtype path, so upcast explicitly.
+        student_logits_chunk = student_input_chunk.float() @ student_weight.float().t()
+        teacher_logits_chunk = teacher_input_chunk.float() @ teacher_weight.float().t()
         chunk_n_rows = student_logits_chunk.shape[0]
 
         # unreduced loss
